flask/requirements.txt
Flask
mysql-connector
pika